feat: bootstrapable themis: split out the bootstrap logic from themis' main package
