+++ .gitignore
@@ -1,5 +1,5 @@
 # Ignore build product.
-themis
+cmd/themis/themis
 
 bin/
 pkg/
