+++ token/claimBuilder.go
@@ -120,10 +120,13 @@ func (nc nonceClaimBuilder) AddClaims(_ context.Context, r *Request, target map[
 
 // remoteClaimBuilder invokes a remote system to obtain claims.
 type remoteClaimBuilder struct {
-	endpoint    endpoint.Endpoint
-	extra       map[string]interface{}
-	apiResults  *prometheus.CounterVec
-	apiDuration prometheus.ObserverVec
+	endpoint      endpoint.Endpoint
+	roots         *x509.CertPool
+	intermediates *x509.CertPool
+	trust         Trust
+	extra         map[string]interface{}
+	apiResults    *prometheus.CounterVec
+	apiDuration   prometheus.ObserverVec
 }
 
 func (rc *remoteClaimBuilder) AddClaims(ctx context.Context, r *Request, target map[string]interface{}) error {
@@ -133,6 +136,7 @@ func (rc *remoteClaimBuilder) AddClaims(ctx context.Context, r *Request, target 
 	maps.Copy(rCopy.PathWildCards, r.PathWildCards)
 	maps.Copy(rCopy.QueryParameters, r.QueryParameters)
 	startTime := time.Now()
+	ctx = SetConnectionDetails(ctx, tlsDetails{TLS: *r.TLS, Roots: rc.roots, Intermediates: rc.intermediates, Trust: rc.trust})
 	result, err := rc.endpoint(sallust.With(ctx, r.Logger), rCopy)
 	duration := time.Since(startTime).Seconds()
 	respErr := RemoteClaimsResponseError{}
@@ -248,7 +252,8 @@ func newRemoteEndpoint(client xhttpclient.Interface, r *RemoteClaims) (endpoint.
 	).Endpoint(), nil
 }
 
-func newRemoteClaimBuilder(endpoint endpoint.Endpoint, metadata map[string]any, r *RemoteClaims, apiResults *prometheus.CounterVec, duration prometheus.ObserverVec) (*remoteClaimBuilder, error) {
+func newRemoteClaimBuilder(endpoint endpoint.Endpoint, metadata map[string]any, roots *x509.CertPool, intermediates *x509.CertPool, trust Trust, r *RemoteClaims, apiResults *prometheus.CounterVec, duration prometheus.ObserverVec) (*remoteClaimBuilder, error) {
+	method := "internal_call"
 	if endpoint == nil {
 		var err error
 
@@ -256,16 +261,16 @@ func newRemoteClaimBuilder(endpoint endpoint.Endpoint, metadata map[string]any, 
 		if err != nil {
 			return nil, err
 		}
-	}
 
-	method := r.Method
-	if len(method) == 0 {
-		method = http.MethodPost
+		method = r.Method
+		if len(method) == 0 {
+			method = http.MethodPost
+		}
 	}
 
 	ls := prometheus.Labels{EndpointLabelKey: r.URL, MethodLabelKey: method}
 
-	return &remoteClaimBuilder{endpoint: endpoint, extra: metadata, apiResults: apiResults.MustCurryWith(ls), apiDuration: duration.MustCurryWith(ls)}, nil
+	return &remoteClaimBuilder{endpoint: endpoint, roots: roots, intermediates: intermediates, trust: trust, extra: metadata, apiResults: apiResults.MustCurryWith(ls), apiDuration: duration.MustCurryWith(ls)}, nil
 }
 
 // newClientCertificateClaimBuiler creates a claim builder that sets trust based
@@ -399,13 +404,13 @@ func NewClaimBuilders(n random.Noncer, remoteEndpoint endpoint.Endpoint, o Optio
 		)
 	}
 
-	if o.Remote != nil {
+	if o.Remote != nil || remoteEndpoint != nil {
 		metadata, err := getStaticValues(o.Metadata)
 		if err != nil {
 			return nil, fmt.Errorf("remote claim builder configuration failure: metadata error: %w", err)
 		}
 
-		remoteClaimBuilder, err := newRemoteClaimBuilder(remoteEndpoint, metadata, o.Remote, remoteResults, remoteDuration)
+		remoteClaimBuilder, err := newRemoteClaimBuilder(remoteEndpoint, metadata, cb.roots, cb.intermediates, cb.trust, o.Remote, remoteResults, remoteDuration)
 		if err != nil {
 			return nil, err
 		}
